[dumpling] tidb: set session variable when load from global variable.
When autocommit is not set by session, we use global variable.
Currently it is loaded in every statement, which means we execute a select statement for every statements.
The behaviour not only hurts performance, but also is wrong.
In MySQL, autocommit is determined when session is created, and it will not be affected when global variable updates.

Also added a basic benchmark method to test basic performace test.