[dumpling] domain: lock reload and fix version check.
 In mysql test , sometimes the domain’s schema version  may be changed
lesser because of another slow transaction.
So we must use lock to guarantee reload once at same time and update
the version check way.
It is very hard to write a test for this, so if we can pass mysql test
using HBase with no error, I think it is fixed.