[dumpling] tidb: remove RestrictedSQLExecutorKey variable in context.
This variable was intended to prevent restricted SQL from being committed.
But it is executed at statement level, will never be committed, so this is not needed.

Also, when begin a new transaction in executing restricted SQL,
If 'autocommit' is 'on'. This variable makes the new transaction into a unwanted `InTrans` state,
so following statements will not be autocommited.