server: update binary protocol for MySQL 5.7
sysbench 1.0.1 failed to run `oltp_read_only.lua` because binary protocal has updated in MySQL 5.7.

About adding test case, this is very expensive to build an synthetic test for this bug.

Because it can only be verified by MySQL client 5.7.

As we strictly follow the new documentation, it's very unlikely this bug can have regression in the future.

So test case is not added.