ddl: remove index_test.go
The tests in `index_test.go` is the old implementation to test index change.
In the tests, we mock DML queries during DDL operation by calling Table functions directly
It's not the same behavior as normal query.

Since we already covered the DDL index change in `index_change_test.go` and this test is unstable.
We can safely remove it.