ddl, util/types: make converter correctly convert string to hex or bit.
This PR fix #2858 and #2717. Furthermore fix the error that
unable to set default value as a hex value just like what happend in #2858.
I also noticed that TiDB have some differences when compared to MySQL.

create table t1 (a bit(64));
insert t1 value ('10');
select a+0 from t1;

In MySQL we can get:
+-------+
| a+0   |
+-------+
| 12592 |
+-------+

But in TiDB we will get:
+-------+
| a+0   |
+-------+
| 10    |
+-------+

As a result of fix #2717, TiDB now behave same as MySQL.