*: fix timestamp column data and index inconsistent involving timezone
To tell the truth, there are many things involved in the bug:
1. Encode/Decode kv range in distsql does not consider timezone
2. DDL background job backfill index always use UTC

I find a clever fix, here it is:
Add a TimeZone field to Time struct, if it's from TiKV, set TimeZone to UTC,
if it's converted from TiDB input, set TimeZone to session's TimeZone.
codec would take TimeZone information when it encoding a timestamp value.