plan: change the value of `descScanFactor` (#6812)
change the value of descScanFactor to make it compatible with tikv due to pingcap/tikv#3124 .
Now, if each row has only one version, the reverse scan's performance is near the performance of the speed of scan. If there's a lot of version, it will become about 1.8x than scan.
So we decrease the factor from 5 to 2.