*: remove `kv.BypassLatch` option and enable latch scheduler by default (#7882)
kv.BypassLatch was introduced to handle the fake transaction confliction problem
in the old implementation. After redesign of the latch scheduler, it's not
needed any more.

Enable latch scheduler makes CI to run it, we can fix bugs and make this feature
more stable.
There is a small performance penalty, less than 5%, users are free to disable it.