store,tidb-server: add a morbid flag to indicate tidb server is closing
When tikv is closed first, and Ctrl+C to tidb then, the server may pending for a while.
Because ddl worker waiting for response, and tikv client retrying the request.
tikv client needs retry to handle network error, but in this case we want it to exit ASAP
Add a morbid flag for that.