[dumpling] expression: add builtin function `WEIGHT_STRING()` (#14792)
Add builtin-function WEIGHT_STRING(). MySQL manual described most of its behavior except:

If the input str is of numeric types, WEIGHT_STRING() in MySQL returns NULL, so does this implementation.
If the input str is of other types rather than string or numeric, the result of WEIGHT_STRING() in MySQL is not cleared. I simply evaluated the input as string.