server: add shutdown, refresh, resetconnection RPCs (#18893)
* server: add shutdown and refresh RPCs

There are several legacy RPC commands which TiDB does not support. This reorganizes the main dispatcher to follow the same order as the protocol to make it clearer which ones are missing.

Support for ComRefresh (often used with mysqldump -F to rotate binary logs) and ComShutdown (legacy, but may still be used by an old client) are added.

I attempted to add ComStatistics, but the RPC is requires a different response packet. So I am going to skip for now.

* Add com_reset_connection rpc

* Add more tests

* Fix reset connection (lost privileges, current db)

Co-authored-by: ti-srebot <66930949+ti-srebot@users.noreply.github.com>