executioner: Fix rouding issues for float and custom colWidth
This resolve issue #22791: query result for float type is incompatible with previous version

It seems like this broke in the fix for issue #21692. In mysql, e-format
is only returned when used in the query, for example:
select 1e15, will print 1e15 back.
However, when stored in a table mysel will print it in decimal form.
issue #21692 resolved this by checking for empty table. However, that
fix applies for everything rather than only for e-format. And that
caused the new issue #22791.

This resolves both #22791 and #21692.

It also resolve an issue where for a default width float column we do
not round to the mysql default precision.