*: query failed after add index / timestamp out-of-range (#28424)
Two issues, convertToMysqlTimestamp did set out-of-range timestamp to
Zero, when it would still be a storable/comparable KindMysqlTime.
(we only artificially limit the range of Timestamp to be compatible
with MySQL).

When this change was done,
it broke #26584 (insert of invalid timestamp succeeded),
which needed a different change in handleZeroDatetime