WIP before cleanup.
We still need to send the ExtraPhysTblID column request to
the store, otherwise there will be issues with the chunk handling.

Like it needs to have space for the column, but the select result
from the store cannot fill it in.

To be cleaned up. This was the last try to avoid sending and
recieve more data from the store for static partition prune mode.

But as discussed, this is a minor issue and is acceptable, due to
static prune mode is planned to be deprecated and removed.