planner: add logic transformation rule groupby_str2int
This rule tries to convert an Aggregation on a string column to Aggregation on int column.
There are several requirements that have to be met so that this rule can be applied:
1. There is a Selection under Aggregation
2. The Selection has equality filter on one of the group keys
3. The filtered group key must be string type
4. The filter condition must be disjoint of IN or EQ scalar function

e.g.
 SELECT a, b, max(c) FROM foo WHERE a in ('x', 'y', 'z') GROUP BY a,b;

will be transformed into
 SELECT CASE WHEN a=0 THEN 'x' WHEN a=1 THEN 'y' WHEN a=2 THEN 'z' END as a, b, max
 FROM (
	SELECT a, b, max(c)
	FROM (
	   SELECT CASE WHEN a='x' THEN 0 WHEN a='y' THEN 1 WHEN a='z' THEN 2 END as a, b, c
	   FROM foo
	   WHERE a in ('x', 'y', 'z')
	) t GROUP BY a,b
 ) t;

NB: This rule doesn't guarantee to yield better alternative, especially when
 1. The IN list constant values are too many, hence the comparison may take a lot more time.
 2. There is an index on the string group key that can be utilized by top Aggregation, whereas
    the alternative will break the order property of the index.

For the moment, we don't consider other special cases, such as the filter condition is below
the child of Aggregation, unless we can pull up the filter from child operators.

Ideally, the optimization should be done in query executor automatically, without intervention
of optimizer, which requires that there is dictionary encoding in the storage engine and the
execution engine can do the transformation on the fly.

This rule is turned off by default. Use with caution.