planner: push down top selection to TiFlash
Before this patch, the planner only push down Selection to TiFlash when it is
required by its parent operator. If the Selection is the top operator of the
plan tree, the planner will failed to push down the Selection even if it is
able to. This patch improves the plan by pushing down Selection whenever
possible.

Close #34192