planner: fix bug that aggregate push down may generate wrong plan for outer joins
When we push down aggregate past outer join, the top count aggregate assumed
that the partial count aggregate always return non-null value. But in case of
outer joins when the aggregate is pushed down to the null generating side of
the outer join, the input for the final aggregate may be nullable.

Before this patch, the final count aggregate always assume its input is not
nullable, hence during aggregate elimimation, the optimizer rewrote aggregate
into a plain column of null-generating side of outer join, which led to the
wrong result.

Close #34465