planner: firstrow agg func should infer return type as nullable for EQAll special case
Before this patch, the fisrtrow agg function infer the return type same with
the argument data type, which is wrong when the arg is not null during the
process planner decorrelate EQAll subqueries, and may cause error on TiFlash
engine.

Currently, firstrow agg function is treated like the exact representation of
aggregate group key, so the data type is the same with group key, even if the
group key is not null.

However, the return type of firstrow should be nullable, we clear the null flag
here instead of during invoking NewAggFuncDesc, in order to keep compatibility
with the existing presumption that the return type firstrow does not change
nullability, whatsoever.

This also can be testified by commit ad7102cd, in which it stated that:
```
For all the aggregate functions except `first_row`, if we have an empty table
defined as t(a,b), `select agg(a) from t` would always return one row, while
`select agg(a) from t group by b` would return empty. For `first_row` which is
only used internally by tidb, `first_row(a)` would always return empty for
empty input now.
```

Close #34584