planner: add a session variable "tidb_opt_inline_cte"
This pr adds a new session variable,
which is mainly used to control whether all CTEs in the entire session are inlined or not.
The default value is false, which means that inline cte is not enabled by default.
(But if the user directly specifies the merge hint, it can still be turned on)
If the variable is set to true, it means that all CTEs for this session try to enable inlining.

Fixed #36514