planner: fix ```can't find column``` for count_star_rewriter
If the index and `count_star` optimizations exist at the same time,
a `can't find column` error may occur.
The main reason is that the `count_star` optimization strategy selects a not null column,
but due to the column is not added to the `colsRequiringFullLen` attribute,
the planner mistakenly selects the index scan.
Eventually the previously selected column could not be found.

This PR fixed #39506 the problem above