Refactor statistics package
This commit refactors the statistics package by removing the statsLogger
variable and replacing it with the StatsLogger variable in the util
package. This change ensures that all logging related to statistics is
consistent and centralized. Additionally, this commit updates the
dependencies in the util package to include the zap logger.

Signed-off-by: hi-rustin <rustin.liu@gmail.com>